third_party/WebKit/LayoutTests/http/tests/inspector/persistence/persistence-do-not-bind-dirty-sourcecode-expected.txt

Verify that dirty uiSourceCodes are not bound.


Running: waitForUISourceCodes

Running: addFileSystemMapping
Failed to create binding: {
       network: http://127.0.0.1:8000/inspector/persistence/resources/foo.js
    fileSystem: file:///var/www/inspector/persistence/resources/foo.js
    exactMatch: false
}
